feat: create shared module
