+++ .github/workflows/ng-test.yml
@@ -26,7 +26,7 @@ jobs:
       - name: Run lint and tests
         run: |
           npm i
-          npm run pr-ci
+          npm run test-ci
       - name: SonarCloud Scan
         uses: SonarSource/sonarcloud-github-action@master
         if: always()
