+++ ng-landing/src/app/about-us/about-us.component.ts
@@ -1,10 +1,10 @@
 import { BehaviorSubject, map } from 'rxjs';
 import { first, tap } from 'rxjs/operators';
 import { Component, OnInit } from '@angular/core';
-import { PageContents } from '../models';
+import { PageContents } from '@shared/models';
 import {
   MonitoringService, PageContentService,
-} from '../services';
+} from '@shared/services';
 
 @Component({
   selector: 'app-about-us',
