--- ng-landing/src/app/models.ts
@@ -1,4 +1,0 @@
-export * from './models/division-content';
-export * from './models/page-contents';
-export * from './models/page-division';
-export * from './models/result';
