--- ng-landing/src/app/models/division-content.ts
@@ -1,6 +1,0 @@
-export interface DivisionContent {
-  id?: string;
-  type: 'paragraph' | undefined;
-  class: string;
-  content: string;
-}
