--- ng-landing/src/app/models/result.ts
@@ -1,4 +1,0 @@
-export interface Result<T> {
-  success: boolean;
-  data: T;
-}
