--- ng-landing/src/app/services.ts
@@ -1,4 +1,0 @@
-export * from './services/api.service';
-export * from './services/app-config.service';
-export * from './services/monitoring.service';
-export * from './services/page-content.service';
