+++ ng-landing/src/app/shared/models.ts
@@ -1,0 +1,4 @@
+export * from './models/division-content';
+export * from './models/page-contents';
+export * from './models/page-division';
+export * from './models/result';
