+++ ng-landing/src/app/shared/models/division-content.ts
@@ -1,0 +1,6 @@
+export interface DivisionContent {
+  id?: string;
+  type: 'paragraph' | undefined;
+  class: string;
+  content: string;
+}
