+++ ng-landing/src/app/shared/models/page-contents.ts
