+++ ng-landing/src/app/shared/models/page-division.ts
