+++ ng-landing/src/app/shared/models/result.ts
@@ -1,0 +1,4 @@
+export interface Result<T> {
+  success: boolean;
+  data: T;
+}
