+++ ng-landing/src/app/shared/services.ts
@@ -1,0 +1,4 @@
+export * from './services/api.service';
+export * from './services/app-config.service';
+export * from './services/monitoring.service';
+export * from './services/page-content.service';
