+++ ng-landing/src/app/shared/services/app-config.service.ts
