+++ ng-landing/src/app/shared/services/page-content.service.ts
